changed to f64 tuple and color

+++ src/color.rs
@@ -9,9 +9,9 @@ use std::ops::{Add, Mul, Sub};
 
 #[derive(Debug, Clone)]
 pub struct Color {
-    red: f32,
-    green: f32,
-    blue: f32,
+    red: f64,
+    green: f64,
+    blue: f64,
 }
 
 // implement '+' operator overload for Colors
@@ -84,11 +84,11 @@ impl Mul<Color> for Color {
     }
 }
 
-// implement '*' operator overload for Colors with a f32
-impl Mul<f32> for Color {
+// implement '*' operator overload for Colors with a f64
+impl Mul<f64> for Color {
     type Output = Color;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Color::new(self.red * rhs, self.green * rhs, self.blue * rhs)
     }
 }
@@ -96,7 +96,7 @@ impl Mul<f32> for Color {
 // Color struct functions
 impl Color {
     // create new color
-    pub fn new(red: f32, green: f32, blue: f32) -> Self {
+    pub fn new(red: f64, green: f64, blue: f64) -> Self {
         Self { red, green, blue }
     }
